contact page is done and header on all pages
