+++ src/app/componentss/carousel/carousel.js
@@ -1,77 +1,71 @@
 "use client";
-import "../../globals.css";
+import { useEffect, useState } from "react";
 
-import { useState, useEffect } from "react";
-
-const images = [
-  "https://placehold.co/1000x50/FF5733/FFF?text=Image+1",
-  "https://placehold.co/1000x50/33FF57/FFF?text=Image+2",
-  "https://placehold.co/1000x50/3357FF/FFF?text=Image+3",
-];
-
-const Carousel = () => {
+function Carousel({ images }) {
   const [currentIndex, setCurrentIndex] = useState(0);
-
-  // Function to go to the next slide
-  const nextSlide = () => {
-    setCurrentIndex((prev) => (prev === images.length - 1 ? 0 : prev + 1));
-  };
-
-  // Function to go to the previous slide
-  const prevSlide = () => {
-    setCurrentIndex((prev) => (prev === 0 ? images.length - 1 : prev - 1));
-  };
-
   // Auto-slide every 5 seconds
   useEffect(() => {
-    const interval = setInterval(nextSlide, 5000);
+    const interval = setInterval(() => {
+      setCurrentIndex((prev) => (prev === images.length - 1 ? 0 : prev + 1));
+    }, 5000);
     return () => clearInterval(interval);
   }, []);
-
   return (
-    <div className="relative w-full h-20 overflow-hidden">
-      {/* Image Slides */}
-      {images.map((src, index) => (
-        <img
-          key={index}
-          src={src}
-          alt={`Slide ${index + 1}`}
-          className={`absolute inset-0 w-full h-20 object-cover transition-opacity duration-700 ${
-            index === currentIndex ? "opacity-100" : "opacity-0"
-          }`}
-        />
-      ))}
-
-      {/* Previous Button */}
-      <button
-        onClick={prevSlide}
-        className="absolute left-4 top-1/2 transform -translate-y-1/2 bg-black bg-opacity-50 text-white p-2 rounded-full hover:bg-opacity-75 transition"
-      >
-        ❮
-      </button>
-
-      {/* Next Button */}
-      <button
-        onClick={nextSlide}
-        className="absolute right-4 top-1/2 transform -translate-y-1/2 bg-black bg-opacity-50 text-white p-2 rounded-full hover:bg-opacity-75 transition"
-      >
-        ❯
-      </button>
-
-      {/* Dots Indicator */}
-      <div className="absolute bottom-2 left-1/2 transform -translate-x-1/2 flex space-x-2">
-        {images.map((_, index) => (
-          <button
+    <div>
+      {" "}
+      {/* Carousel */}
+      <div className="relative w-full h-20 overflow-hidden">
+        {/* Images */}
+        {images.map((src, index) => (
+          <img
             key={index}
-            onClick={() => setCurrentIndex(index)}
-            className={`w-3 h-3 rounded-full ${
-              index === currentIndex ? "bg-white" : "bg-gray-400"
+            src={src}
+            alt={`Slide ${index + 1}`}
+            className={`absolute inset-0 w-full h-20 object-cover transition-opacity duration-700 ${
+              index === currentIndex ? "opacity-100" : "opacity-0"
             }`}
           />
         ))}
+
+        {/* Previous Button */}
+        <button
+          onClick={() =>
+            setCurrentIndex((prev) =>
+              prev === 0 ? images.length - 1 : prev - 1
+            )
+          }
+          className="absolute left-4 top-1/2 transform -translate-y-1/2 bg-black bg-opacity-50 text-white p-2 rounded-full hover:bg-opacity-75 transition"
+        >
+          ❮
+        </button>
+
+        {/* Next Button */}
+        <button
+          onClick={() =>
+            setCurrentIndex((prev) =>
+              prev === images.length - 1 ? 0 : prev + 1
+            )
+          }
+          className="absolute right-4 top-1/2 transform -translate-y-1/2 bg-black bg-opacity-50 text-white p-2 rounded-full hover:bg-opacity-75 transition"
+        >
+          ❯
+        </button>
+
+        {/* Dots Indicator */}
+        <div className="absolute bottom-2 left-1/2 transform -translate-x-1/2 flex space-x-2">
+          {images.map((_, index) => (
+            <button
+              key={index}
+              onClick={() => setCurrentIndex(index)}
+              className={`w-3 h-3 rounded-full ${
+                index === currentIndex ? "bg-white" : "bg-gray-400"
+              }`}
+            />
+          ))}
+        </div>
       </div>
     </div>
   );
-};
+}
 
 export default Carousel;
