+++ src/app/page.js
@@ -1,13 +1,13 @@
 import "./globals.css";
-import Header from "./componentss/header/page";
+import Header from "./componentss/header/header";
 import Main from "./componentss/main/page";
 import Carousel from "./componentss/carousel/carousel";
 export default function Home() {
   return (
     <main>
-       <Header />
-       {/* <Main /> */}
-       {/* <Carousel /> */}
+      <Header />
+      {/* <Main /> */}
+      {/* <Carousel /> */}
     </main>
   );
 }
