+++ src/app/position/page.js
@@ -1,20 +1,46 @@
-import React from 'react'
+"use client";
+import { CircleChevronDown, CircleChevronRight } from "lucide-react";
+import React, { useState } from "react";
 
 function Page() {
+  const [clicked, setClicked] = useState(false);
   return (
-
-<>
-<div className='flex items-center h-screen'>
-<div className='bg-gray-600 w-full h-[200px] absolute'>
-
-    <div className=' w-[50px] h-[50px] bg-red-300 rounded-[50%] absolute bottom-[50%] right-[50%]'>
-
-</div>
-
-    </div>
-    </div>
-</>
-  )
+    <>
+      <div className="flex items-center h-screen">
+        <hr />
+        <hr />
+        <hr />
+        <br />
+        <div className="ml-20">
+          <div class="flex">
+            {clicked ? <CircleChevronDown /> : <CircleChevronRight />}{" "}
+            <h3
+              onClick={() => {
+                setClicked(!clicked);
+              }}
+              className="font-semibold"
+            >
+              Click me
+            </h3>
+          </div>
+          <div className={clicked ? "block" : "max-md:hidden"}>
+            <p>Halal</p>
+            <p>Halal</p>
+            <p>Halal</p>
+            <p>Halal</p>
+            <p>Halal</p>
+            <p>Halal</p>
+            <p>Halal</p>
+            <p>Halal</p>
+            <p>Halal</p>
+            <p>Halal</p>
+            <p>Halal</p>
+            <p>Halal</p>
+          </div>
+        </div>
+      </div>
+    </>
+  );
 }
 
-export default Page
+export default Page;
